requirements.txt
amqp==2.3.2
anyjson==0.3.3
async==0.6.2
billiard==3.5.0.5
celery==4.2.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.2
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-timezone-field==3.0
django-widget-tweaks==1.4.3
dnspython==1.16.0
eventlet==0.24.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
kombu==4.2.2
Markdown==3.0.1
monotonic==1.5
Pillow==5.3.0
psycopg2-binary==2.7.6.1
pycparser==2.19
python-crontab==2.3.5
python-dateutil==1.5
pytz==2018.7
rabbitmq==0.2.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
vine==1.1.4
whitenoise==4.1.2
wincertstore==0.2
